update from ec2

+++ utilities/distributed_utilities.py
@@ -3,6 +3,7 @@ import subprocess
 import boto3
 import os
 import sys
+import time
 import cPickle as pickle
 import json
 from utilities2015 import execute_command
@@ -11,20 +12,26 @@ from metadata import *
 def delete_file_or_directory(fp):
     execute_command("rm -rf %s" % fp)
 
-def transfer_data(from_fp, to_fp, from_hostname, to_hostname, is_dir, include_only=None):
+def transfer_data(from_fp, to_fp, from_hostname, to_hostname, is_dir, include_only=None, exclude_only=None, includes=None):
     assert from_hostname in ['localhost', 'workstation', 'oasis', 's3', 'ec2', 's3raw'], 'from_hostname must be one of localhost, workstation, oasis, s3, s3raw or ec2.'
     assert to_hostname in ['localhost', 'workstation', 'oasis', 's3', 'ec2', 's3raw'], 'to_hostname must be one of localhost, workstation, oasis, s3, s3raw or ec2.'
 
     to_parent = os.path.dirname(to_fp)
     
     #oasis = 'oasis-dm.sdsc.edu'
+    
+    t = time.time()
 
     if from_hostname in ['localhost', 'ec2', 'workstation']:
         # upload
         if to_hostname in ['s3', 's3raw']:
             if is_dir:
-                if include_only is not None:
+                if includes is not None:
+                    execute_command('aws s3 cp --recursive %(from_fp)s s3://%(to_fp)s --exclude \"*\" %(includes_str)s' % dict(from_fp=from_fp, to_fp=to_fp, includes_str=" ".join(['--include ' + incl for incl in includes])))
+                elif include_only is not None:
                     execute_command('aws s3 cp --recursive %(from_fp)s s3://%(to_fp)s --exclude \"*\" --include \"%(include)s\"' % dict(from_fp=from_fp, to_fp=to_fp, include=include_only))
+                elif exclude_only is not None:
+                    execute_command('aws s3 cp --recursive %(from_fp)s s3://%(to_fp)s --include \"*\" --exclude \"%(exclude)s\"' % dict(from_fp=from_fp, to_fp=to_fp, exclude=exclude_only))
                 else:
                     execute_command('aws s3 cp --recursive %(from_fp)s s3://%(to_fp)s' % \
             dict(from_fp=from_fp, to_fp=to_fp))
@@ -39,13 +46,17 @@ def transfer_data(from_fp, to_fp, from_hostname, to_hostname, is_dir, include_on
         if from_hostname in ['s3', 's3raw']:
 
             # Clear existing folder/file
-            if not include_only:
+            if not include_only and not includes and not exclude_only:
                 execute_command('rm -rf %(to_fp)s && mkdir -p %(to_parent)s' % dict(to_parent=to_parent, to_fp=to_fp))
 
             # Download from S3 using aws commandline interface.
             if is_dir:
-                if include_only is not None:
+                if includes is not None:
+                    execute_command('aws s3 cp --recursive s3://%(from_fp)s %(to_fp)s --exclude \"*\" %(includes_str)s' % dict(from_fp=from_fp, to_fp=to_fp, includes_str=" ".join(['--include ' + incl for incl in includes])))
+                elif include_only is not None:
                     execute_command('aws s3 cp --recursive s3://%(from_fp)s %(to_fp)s --exclude \"*\" --include \"%(include)s\"' % dict(from_fp=from_fp, to_fp=to_fp, include=include_only))
+                elif exclude_only is not None:
+                    execute_command('aws s3 cp --recursive s3://%(from_fp)s %(to_fp)s --include \"*\" --exclude \"%(exclude)s\"' % dict(from_fp=from_fp, to_fp=to_fp, exclude=exclude_only))
                 else:
                     execute_command('aws s3 cp --recursive s3://%(from_fp)s %(to_fp)s' % dict(from_fp=from_fp, to_fp=to_fp))
             else:
@@ -56,10 +67,13 @@ def transfer_data(from_fp, to_fp, from_hostname, to_hostname, is_dir, include_on
         # log onto another machine and perform upload from there.
         execute_command("ssh %(from_hostname)s \"ssh %(to_hostname)s \'rm -rf %(to_fp)s && mkdir -p %(to_parent)s && scp -r %(from_fp)s %(to_hostname)s:%(to_fp)s\'\"" % \
                         dict(from_fp=from_fp, to_fp=to_fp, from_hostname=from_hostname, to_hostname=to_hostname, to_parent=to_parent))
+    
+    sys.stderr.write('%.2f seconds.\n' % (time.time() - t))
+        
 
 default_root = dict(localhost='/home/yuncong',workstation='/media/yuncong/BstemAtlasData', oasis='/home/yuncong/csd395', s3=S3_DATA_BUCKET, ec2='/shared', s3raw=S3_RAWDATA_BUCKET)
 
-def transfer_data_synced(fp_relative, from_hostname, to_hostname, is_dir, from_root=None, to_root=None, include_only=None, s3_bucket=None):
+def transfer_data_synced(fp_relative, from_hostname, to_hostname, is_dir, from_root=None, to_root=None, include_only=None, exclude_only=None, includes=None, s3_bucket=None):    
     if from_root is None:
         from_root = default_root[from_hostname]
     if to_root is None:
@@ -67,7 +81,8 @@ def transfer_data_synced(fp_relative, from_hostname, to_hostname, is_dir, from_r
 
     from_fp = os.path.join(from_root, fp_relative)
     to_fp = os.path.join(to_root, fp_relative)
-    transfer_data(from_fp=from_fp, to_fp=to_fp, from_hostname=from_hostname, to_hostname=to_hostname, is_dir=is_dir, include_only=include_only)
+    transfer_data(from_fp=from_fp, to_fp=to_fp, from_hostname=from_hostname, to_hostname=to_hostname, is_dir=is_dir, include_only=include_only, exclude_only=exclude_only, includes=includes)
+
 
 def first_last_tuples_distribute_over(first_sec, last_sec, n_host):
     secs_per_job = (last_sec - first_sec + 1)/float(n_host)
@@ -134,14 +149,14 @@ def detect_responsive_nodes(exclude_nodes=[], use_nodes=None):
 
     return up_hostids
 
-def run_distributed(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclude_nodes=[], use_nodes=None, argument_type='list', cluster_size=None):
+def run_distributed(command, kwargs_list, stdout=open('/tmp/log', 'ab+'), exclude_nodes=[], use_nodes=None, argument_type='list', cluster_size=None, jobs_per_node=None):
     if ON_AWS:
-        run_distributed5(command, kwargs_list, cluster_size, stdout, argument_type)
+        run_distributed5(command=command, kwargs_list=kwargs_list, cluster_size=cluster_size, jobs_per_node=jobs_per_node, stdout=stdout, argument_type=argument_type)
     else:
         run_distributed4(command, kwargs_list, stdout, exclude_nodes, use_nodes, argument_type)
 
 
-def run_distributed5(command, kwargs_list, cluster_size, stdout=open('/tmp/log', 'ab+'), argument_type='list'):
+def run_distributed5(command, kwargs_list, cluster_size, jobs_per_node=1, stdout=open('/tmp/log', 'ab+'), argument_type='list'):
     """
     Distributed executing a command on AWS.
     """
@@ -159,11 +174,11 @@ def run_distributed5(command, kwargs_list, cluster_size, stdout=open('/tmp/log',
         subprocess.call("aws autoscaling set-desired-capacity --auto-scaling-group-name %s --desired-capacity %d" % (asg, cluster_size), shell=True)
         print "Setting autoscaling group %s capaticy to %d." % (asg, cluster_size)
 
-        print "Wait for SGE to know all nodes..."
-
-        # Wait for SGE to know all nodes. Timeout = 10 mins
+        # Wait for SGE to know all nodes.
+        create_fleet_wait_seconds = 60
+        print "Wait for SGE to know all nodes (timeout in %d seconds)..." % create_fleet_wait_seconds
         success = False
-        for _ in range(600/5):
+        for _ in range(create_fleet_wait_seconds/5):
             n_hosts = (subprocess.check_output('qhost')).count('\n') - 3
             if n_hosts == cluster_size:
                 success = True
@@ -171,11 +186,11 @@ def run_distributed5(command, kwargs_list, cluster_size, stdout=open('/tmp/log',
             time.sleep(5)
 
         if not success:
-            raise Exception('SGE does not receive all host information in 300 seconds. Abort.')
+            sys.stderr.write('SGE does not receive all host information in %d seconds. Continue with the %d nodes currently available.' % (create_fleet_wait_seconds, n_hosts))
         else:
-            print "All nodes are ready."
+            sys,stderr.write("All nodes are ready.\n")
 
-    assert n_hosts >= cluster_size
+    # assert n_hosts >= cluster_size
 
     temp_script = '/tmp/runall.sh'
 
@@ -212,9 +227,14 @@ def run_distributed5(command, kwargs_list, cluster_size, stdout=open('/tmp/log',
         temp_f.write(line)
         temp_f.close()
         os.chmod(temp_script, 0o777)
-        call('qsub -V -l mem_free=60G -o %(stdout_log)s -e %(stderr_log)s %(script)s' % \
-             dict(script=temp_script, stdout_log='/home/ubuntu/stdout_%d.log' % i, stderr_log='/home/ubuntu/stderr_%d.log' % i),
+        # call('qsub -V -l mem_free=60G -o %(stdout_log)s -e %(stderr_log)s %(script)s' % \
+        #      dict(script=temp_script, stdout_log='/home/ubuntu/stdout_%d.log' % i, stderr_log='/home/ubuntu/stderr_%d.log' % i),
+        #      shell=True, stdout=stdout)
+        
+        call('qsub -pe mpi %(jobs_per_node)d -V -l mem_free=60G -o %(stdout_log)s -e %(stderr_log)s %(script)s' % \
+             dict(jobs_per_node=jobs_per_node, script=temp_script, stdout_log='/home/ubuntu/stdout_%d.log' % i, stderr_log='/home/ubuntu/stderr_%d.log' % i),
              shell=True, stdout=stdout)
+        
 
     # Wait for qsub to complete.
     success = False
